Unit tests for issue_milestone
