+++ models/fixtures/issue.yml
@@ -19,6 +19,7 @@
   poster_id: 1
   name: issue2
   content: content2
+  milestone_id: 1
   is_closed: false
   is_pull: true
   created_unix: 946684810
