+++ models/fixtures/repository.yml
@@ -8,6 +8,7 @@
   num_closed_issues: 1
   num_pulls: 2
   num_closed_pulls: 0
+  num_milestones: 2
   num_watches: 2
 
 -
